requirements.txt
Django
djangorestframework
channels
